Split classes into separate files
According to the [PSR-1][1] coding standard:

> each class must be in a file by itself

[1]: https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-1-basic-coding-standard.md
